fix further tests
