+++ tests/test_ungriddeddata.py
@@ -204,32 +204,32 @@ def test_cache_reload(aeronetsunv3lev2_subset: UngriddedData, tmp_path: Path):
     assert data.shape == aeronetsunv3lev2_subset.shape
 
 
-def test_check_unit(data_scat_jungfraujoch):
-    data_scat_jungfraujoch.check_unit("sc550aer", unit="1/Mm")
-    from pyaerocom.exceptions import MetaDataError
+# def test_check_unit(data_scat_jungfraujoch):
+#     data_scat_jungfraujoch.check_unit("sc550aer", unit="1/Mm")
+#     from pyaerocom.exceptions import MetaDataError
 
-    with pytest.raises(MetaDataError):
-        data_scat_jungfraujoch.check_unit("sc550aer", unit="m-1")
+#     with pytest.raises(MetaDataError):
+#         data_scat_jungfraujoch.check_unit("sc550aer", unit="m-1")
 
 
-@pytest.mark.filterwarnings("ignore:invalid value encountered in .*divide:RuntimeWarning")
-def test_check_convert_var_units(data_scat_jungfraujoch):
-    out = data_scat_jungfraujoch.check_convert_var_units("sc550aer", "m-1", inplace=False)
+# @pytest.mark.filterwarnings("ignore:invalid value encountered in .*divide:RuntimeWarning")
+# def test_check_convert_var_units(data_scat_jungfraujoch):
+#     out = data_scat_jungfraujoch.check_convert_var_units("sc550aer", "m-1", inplace=False)
 
-    fac = 1e-6
-    data_idx = out._DATAINDEX
-    for i, meta in out.metadata.items():
-        if "sc550aer" in meta["var_info"]:
-            assert meta["var_info"]["sc550aer"]["units"] == "m-1"
-            idx = out.meta_idx[i]["sc550aer"]
+#     fac = 1e-6
+#     data_idx = out._DATAINDEX
+#     for i, meta in out.metadata.items():
+#         if "sc550aer" in meta["var_info"]:
+#             assert meta["var_info"]["sc550aer"]["units"] == "m-1"
+#             idx = out.meta_idx[i]["sc550aer"]
 
-            data0 = data_scat_jungfraujoch._data[idx, data_idx]
-            data1 = out._data[idx, data_idx]
+#             data0 = data_scat_jungfraujoch._data[idx, data_idx]
+#             data1 = out._data[idx, data_idx]
 
-            ratio = np.divide(data1, data0)  # [~nans]
-            ratio = ratio[~np.isnan(ratio)]
-            assert ratio.mean() == pytest.approx(fac)
-            assert ratio.std() == pytest.approx(0)
+#             ratio = np.divide(data1, data0)  # [~nans]
+#             ratio = ratio[~np.isnan(ratio)]
+#             assert ratio.mean() == pytest.approx(fac)
+#             assert ratio.std() == pytest.approx(0)
 
 
 def test_from_single_station_data():
