+++ tests/test_ungriddeddata_structured.py
@@ -193,36 +193,33 @@ def test_coordinate_access():
 #     assert data.shape == aeronetsunv3lev2_subset.shape
 
 
-# def test_check_unit(data_scat_jungfraujoch):
-#     data_scat_jungfraujoch.check_unit("sc550aer", unit="1/Mm")
-#     from pyaerocom.exceptions import MetaDataError
+def test_check_unit(data_scat_jungfraujoch):
+    data_scat_jungfraujoch.check_unit("sc550aer", unit="1/Mm")
+    from pyaerocom.exceptions import MetaDataError
 
-#     with pytest.raises(MetaDataError):
-#         data_scat_jungfraujoch.check_unit("sc550aer", unit="m-1")
+    with pytest.raises(MetaDataError):
+        data_scat_jungfraujoch.check_unit("sc550aer", unit="m-1")
 
 
-# @pytest.mark.filterwarnings(
-#     "ignore:invalid value encountered in .*divide:RuntimeWarning"
-# )
-# def test_check_convert_var_units(data_scat_jungfraujoch):
-#     out = data_scat_jungfraujoch.check_convert_var_units(
-#         "sc550aer", "m-1", inplace=False
-#     )
+@pytest.mark.filterwarnings("ignore:invalid value encountered in .*divide:RuntimeWarning")
+def test_check_convert_var_units(data_scat_jungfraujoch):
+    out = data_scat_jungfraujoch.check_convert_var_units("sc550aer", "m-1", inplace=False)
 
-#     fac = 1e-6
-#     data_idx = out._DATAINDEX
-#     for i, meta in out.metadata.items():
-#         if "sc550aer" in meta["var_info"]:
-#             assert meta["var_info"]["sc550aer"]["units"] == "m-1"
-#             idx = out.meta_idx[i]["sc550aer"]
+    fac = 1e-6
+    for i, meta in out.metadata.items():
+        if "sc550aer" in meta["var_info"]:
+            assert meta["var_info"]["sc550aer"]["units"] == "m-1"
+            idx = (out._data.data["meta_id"] == i) & (
+                out._data.data["var_id"] == out.var_idx["sc550aer"]
+            )
 
-#             data0 = data_scat_jungfraujoch._data[idx, data_idx]
-#             data1 = out._data[idx, data_idx]
+            data0 = data_scat_jungfraujoch._data.data["data"][idx]
+            data1 = out._data.data["data"][idx]
 
-#             ratio = np.divide(data1, data0)  # [~nans]
-#             ratio = ratio[~np.isnan(ratio)]
-#             assert ratio.mean() == pytest.approx(fac)
-#             assert ratio.std() == pytest.approx(0)
+            ratio = np.divide(data1, data0)  # [~nans]
+            ratio = ratio[~np.isnan(ratio)]
+            assert ratio.mean() == pytest.approx(fac)
+            assert ratio.std() == pytest.approx(0)
 
 
 def test_from_single_station_data():
